# Add Rect attribute kwargs for Surface.blit 