--- a/src_c/surface.c
+++ b/src_c/surface.c
@@ -1853,7 +1853,7 @@ surf_blit(pgSurfaceObject *self, PyObject *args, PyObject *keywds)
         "bottomleft", "bottomright", "midleft", "midright",      "midtop",
         "midbottom",  "center",      NULL};
     if (!PyArg_ParseTupleAndKeywords(
-            args, keywds, "O!|OOiOOOOOOO", kwids, &pgSurface_Type, &srcobject,
+            args, keywds, "O!|OOiOOOOOOOO", kwids, &pgSurface_Type, &srcobject,
             &topleft_pos, &argrect, &the_args, &topright_pos, &bottomleft_pos,
             &bottomright_pos, &midleft_pos, &midright_pos, &midtop_pos,
             &midbottom_pos, &center_pos))