… docs/reST/ref/surface.rst
… test.py
… test/blit_test.py