Fixes: https://github.com/pygame/pygame/issues/3202