@novialriptide 2022-06-05
> As MyreMylar expressed in the linked issue how this could be a new function, I quite like that suggestion myself too

I wouldn't mind this either.