@novialriptide 2022-06-13
A new PR is in the works right now, it'll either be `pygame.Surface.paste()` or something else... I'll close this PR for now as this will no longer be in use.