@novialriptide 2022-06-28
Reopening since it seems like `.blit()` is favored.