@MyreMylar 2022-07-10
I think sticking this in blit is fine if we are also having raw blit without all the bells and whistles. I don't think it will have that much performance impact anyway (and there is still performance work to be done on the slower alpha blits at the pixel level anyway)