@novialriptide 2022-07-15
I realized that there was a big problem, so I did this little workaround to fix it.

Right now, we think that legacy code could work as such like this (let's say `mario_bulletbill.png` is an existing image):
```python
bulletbill = pygame.image.load("mario_bulletbill.png")
s = pygame.Surface((256, 224))
s.blit(bulletbill, (0, 0), area=None, special_flags=BLEND_SUB)
```
But the problem is that the kwarg values could be inputted like this
```python
bulletbill = pygame.image.load("mario_bulletbill.png")
s = pygame.Surface((256, 224))
s.blit(bulletbill, (0, 0), None, BLEND_SUB)
```

The code will detect `(0, 0)` as the `topleft` kwarg, `None` as the `topright` kwarg, and so on.

I don't think this should be too big of a problem anyways. I'm just pointing it out.