@novialriptide 2022-07-18
## Performance Test
## New code (using this branch)
```python
import pygame
import timeit
surf = pygame.Surface((255, 255))
surf2 = pygame.Surface((10, 10))
surf2.fill((255, 0, 0))

timed = timeit.timeit(lambda: surf.blit(surf2, topright=(0, 0)), number=1)
print(timed, "ms")
```
```shell
pygame 2.1.3.dev5 (SDL 2.0.22, Python 3.9.0)
Hello from the pygame community. https://www.pygame.org/contribute.html
1.5399999999998748e-05 ms
```
## Old code (using this branch)
```python
import pygame
import timeit
surf = pygame.Surface((255, 255))
surf2 = pygame.Surface((10, 10))
surf2.fill((255, 0, 0))

timed = timeit.timeit(lambda: surf.blit(surf2, (0, 0)), number=1)
print(timed, "ms")
```
```shell
pygame 2.1.3.dev5 (SDL 2.0.22, Python 3.9.0)
Hello from the pygame community. https://www.pygame.org/contribute.html
1.580000000001025e-05 ms
```
## Old code (pygame 2.1.2)
```python
import pygame
import timeit
surf = pygame.Surface((255, 255))
surf2 = pygame.Surface((10, 10))
surf2.fill((255, 0, 0))

timed = timeit.timeit(lambda: surf.blit(surf2, (0, 0)), number=1)
print(timed, "ms")
```
```shell
pygame 2.1.2 (SDL 2.0.18, Python 3.9.0)
Hello from the pygame community. https://www.pygame.org/contribute.html
1.1000000000011001e-05 ms
```