@novialriptide 2022-07-18
Here's another performance test
## Performance Test (part 2)
### pygame 2.1.3.dev5 (this branch)
```python
import pygame
import timeit
surf = pygame.Surface((255, 255))
surf2 = pygame.Surface((10, 10))
surf2.fill((255, 0, 0))

timed = timeit.timeit(lambda: surf.blit(surf2, topleft=(100, 100)), number=10000000)
print(f"topleft: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, topright=(100, 100)), number=10000000)
print(f"topright: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, bottomleft=(100, 100)), number=10000000)
print(f"bottomleft: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, bottomright=(100, 100)), number=10000000)
print(f"bottomright: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, midleft=(100, 100)), number=10000000)
print(f"midleft: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, midright=(100, 100)), number=10000000)
print(f"midright: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, midtop=(100, 100)), number=10000000)
print(f"midtop: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, midbottom=(100, 100)), number=10000000)
print(f"midbottom: {timed} seconds")

timed = timeit.timeit(lambda: surf.blit(surf2, center=(100, 100)), number=10000000)
print(f"center: {timed} seconds")
```
topleft: 4.0847878 seconds
topright: 5.235740300000001 seconds
bottomleft: 5.8812159999999984 seconds
bottomright: 6.3172087999999995 seconds
midleft: 6.667041000000001 seconds
midright: 7.170022800000002 seconds
midtop: 7.438604399999996 seconds
midbottom: 7.826056099999995 seconds
center: 8.368447199999999 seconds
```

### pygame 2.1.2
```python
import pygame
import timeit
surf = pygame.Surface((255, 255))
surf2 = pygame.Surface((10, 10))
surf2.fill((255, 0, 0))

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(topleft=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"topleft: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(topright=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"topright: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(bottomleft=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"bottomleft: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(bottomright=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"bottomright: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(midleft=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"midleft: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(midright=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"midright: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(midtop=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"midtop: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(midbottom=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"midbottom: {timed} seconds")

timed = timeit.timeit("surf.blit(surf2, surf2.get_rect(center=(0, 0)))", globals={"surf": surf, "surf2": surf2}, number=10000000)
print(f"center: {timed} seconds")
```

```shell
pygame 2.1.3.dev5 (SDL 2.0.22, Python 3.9.0)
Hello from the pygame community. https://www.pygame.org/contribute.html
topleft: 3.7194558 seconds
topright: 3.1512507000000003 seconds
bottomleft: 3.1497751000000003 seconds
bottomright: 3.168838599999999 seconds
midleft: 3.5747503 seconds
midright: 3.1675736000000008 seconds
midtop: 3.5943006000000004 seconds
midbottom: 3.190688999999999 seconds
center: 3.4581815999999996 seconds
```

## Conclusion
I need to optimize this a bit...